requirements.txt
torch==1.9.1
torchvision==0.10.1
facenet-pytorch==2.5.2
opencv-python==4.5.3.56
numpy==1.21.2
scikit-learn==0.24.2
Pillow==8.3.1
